Moved `AsyncLock` to Core lib
